PiP enhanced
Prepares Picture in Picture implementation for user mouse click. That is, once we can extract the mouse position, modifying this version should result in desired Picture in Picture implementation.

Tracking is done by calculating the distance between the clicked position and the found players. The smallest distance is the one shown in the PiP.